// src/Wrapper/Wrapper.did.d.ts
export type AccountIdentifier = AccountIdentifier_2;
export type AccountIdentifier_2 = string;
export type AccountIdentifier_3 = AccountIdentifier;
// ...
export type Time = Time_2;
export type Time_2 = bigint;
export type TokenIdentifier = TokenIdentifier_2;
export type TokenIdentifier_2 = string;
export type TokenIndex = TokenIndex_2;
export type TokenIndex_2 = number;
export interface Transaction {
  token: TokenIdentifier;
  time: Time;
  seller: Principal;
  buyer: AccountIdentifier_3;
  price: bigint;
}